Mental - but kind of works - see details below
Ended up talking to @manics (how frigging cool is that) -> https://discourse.jupyter.org/t/after-oauthenticator-successfully-returned-the-user-kubespawner-stuck-at-creating-pvc/17056

Turns out there is a problem with OAuthenticator not working with http_proxy. However, the opposite true in my case.

AAD authentication only works when `https_proxy` is set. However, when this env var is set,. subsequent pod deployment fails. I suspsect I have not configured the `no_proxy` variable correctly - but that is for another day.

So, this is what this commit does:

- Replace the `oauth2.py` file in oauthenticator with a custom file. In this do the following:
  - Set `os.environ["https_proxy"] = "http://lthswproxy01.xlthtr.nhs.uk:800"`
  - `AsyncHTTPClient.configure("tornado.curl_httpclient.CurlAsyncHTTPClient")` to make use of proxies.
  - Once authentication is successful, set `os.environ["https_proxy"] =""`
  - Needs tidying up and checking if this works for multiple users and what happens in _race_ conditions when more than one user is trying to log in at the same time.
  -